fix naming,
divided to separate files,

use CamelCase
