+++ CourseApp.Tests/Div4in0returnedError.cs
@@ -3,14 +3,14 @@ namespace TestsDiv
     using CourseApp;
     using Xunit;
 
-    public class Div_4in2_Re2
+    public class Div4in0returnedError
     {
         [Fact]
-        public void Div_4Div2_Returned2()
+        public void DivPosZeroErr()
         {
             double arg1 = 4;
-            double arg2 = 2;
-            double expected = 2;
+            double arg2 = 0;
+            double expected = 0;
 
             // act
             var actual = new Calculator().Division(arg1, arg2);
