--- CourseApp.Tests/FunctionCalculatorTest.cs
@@ -1,14 +1,0 @@
-﻿namespace CourseApp.Tests
-{
-    using Xunit;
-
-    public class FunctionCalculatorTest
-    {
-        [Fact]
-
-        public void Test()
-        {
-            Assert.True(true);
-        }
-    }
-}
