--- CourseApp/Program.cs
@@ -1,13 +1,0 @@
-﻿namespace CourseApp
-{
-    using System;
-
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            Console.WriteLine($"Hello world");
-            Console.ReadLine();
-        }
-    }
-}
